+++ Ruby-Project/person_class.rb
@@ -1,0 +1,26 @@
+class Person
+  attr_accessor :name, :age
+  attr_reader :id
+
+  def initialize(age, name = 'Unknown', parent_permission: true)
+    @id = Random.rand(1..1000)
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+  end
+
+  def can_use_services?
+    true if @age >= 18 || @parent_permission == true
+  end
+
+  private
+
+  def of_age?
+    if @age >= 18
+      @parent_permission = true
+    elsif @age < 18
+      @parent_permission = false
+
+    end
+  end
+end
