--- Ruby-Project/student-class.rb
@@ -1,10 +1,0 @@
-require "./Person-class.ru"
-class Student < Person
-    def initialize(classroom)
-        @classroom = classroom
-    end
-
-    def play_hooky
-        return "¯\(ツ)/¯"
-    end
-end
