+++ Ruby-Project/student_class.rb
@@ -1,0 +1,11 @@
+require './person_class'
+class Student < Person
+  def initialize
+    super(classroom)
+    @classroom = classroom
+  end
+
+  def play_hooky
+    '¯(ツ)/¯'
+  end
+end
