--- Ruby-Project/teacher-class.rb
@@ -1,6 +1,0 @@
-require "./Person-class.ru"
-class Teacher < person
-def initialize(specialization)
-    @specialization = specialization
-    @can_use_services = true
-end
