+++ Ruby-Project/teacher_class.rb
@@ -1,0 +1,8 @@
+require './person_class'
+class Teacher < person
+  def initialize
+    super(@specialization)
+    @specialization = specialization
+    @can_use_services = true
+  end
+end
